requirements.txt
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dnspython==2.2.1
Flask==2.1.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
Werkzeug==2.1.2
zipp==3.8.0
